bookutu-backendz/requirements.txt
# Django framework
Django==4.2.7

# REST API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-filter==23.3
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Task queue
celery==5.3.4
redis==5.0.1

# Image processing
Pillow==10.0.1

# HTTP requests
requests==2.32.0

